requirements.txt
# Mirrors `backend/requirements.txt` — keep in sync for local `pip install -r requirements.txt`.
# See backend/requirements.txt for the canonical list.

groq>=0.4.2
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
gunicorn>=23.0.0
sqlalchemy>=2.0.36
psycopg[binary]>=3.2.0
pydantic>=2.11.0,<3
email-validator>=2.2.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1

httpx>=0.27.0
requests>=2.31.0

beautifulsoup4>=4.12.0
googlesearch-python>=1.2.0

aiosmtplib>=2.0.0

SpeechRecognition==3.10.4
pydub==0.25.1

gTTS==2.5.1
